refactor(use-shopify): restructure src directory
