+++ src/graphql/queries/index.ts
