+++ src/graphql/queries/sharedTypes.ts
