+++ src/types/customer.ts
