+++ src/types/graphql.ts
